+++ CHANGELOG.md
@@ -9,6 +9,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Major
 
 ### Minor
+
+### Patch
+
+## [v2.4.0] - 2017-08-16
+
+### Minor
 - Make `catchify.limit(iterable, limit, exitOnError)` support objects
 - Make `catchify.some(iterable)` support objects
 - Make `catchify.all(iterable)` support objects
@@ -127,7 +133,8 @@ an error does not prevent all errors and values from being returned
 ### Major
 - Initial release
 
-[Unreleased]: https://github.com/majgis/catchify/compare/v2.3.4...master
+[Unreleased]: https://github.com/majgis/catchify/compare/v2.4.0...master
+[v2.4.0]: https://github.com/majgis/catchify/compare/v2.3.4...v2.4.0
 [v2.3.4]: https://github.com/majgis/catchify/compare/v2.3.3...v2.3.4
 [v2.3.3]: https://github.com/majgis/catchify/compare/v2.3.2...v2.3.3
 [v2.3.2]: https://github.com/majgis/catchify/compare/v2.3.1...v2.3.2